fix: unnecessary remap to select-mode which disturbs nvim-snippy
`vmap` is used to map `gx` to visual-mode. However, `vmap` not only maps
to visual-mode but also select-mode. This is first of all not necessary
because `xmap` is enough to map to visual-mode but more importantly it
disturbs other plugins which rely on select-mode like `nvim-snippy`.
During snippet-expansion snippy selects with the select-mode
placeholders which should be overwritten by the user. Typing `g` becomes
impossible without pressing another 2nd key because it's remapped by
gx-extended. It is generally good practice to use xmap instead of vmap
unless necessary.